Practice_1 week 4 commit.1

+++ week3/practice_6/src/main.rs
@@ -1,4 +1,4 @@
 fn main() {
-    let interest:i32 = 8 as f32 ; // integer assigned to float variable
+    let interest:f32 = 8 ; // integer assigned to float variable
     println!("interest is {}",interest );
 }
